+++ src/groups/special_orthogonal_group.jl
@@ -78,12 +78,7 @@ function identity_element(G::SpecialOrthogonalGroup, ::Type{<:AbstractMatrix{T}}
     return identity_element!(G, e)
 end
 
-inv!(G::SpecialOrthogonalGroup, k, g) = copyto!(G, k, transpose(g))
-function inv!(
-        G::SpecialOrthogonalGroup, q, ::Identity{O}
-    ) where {O <: AbstractMultiplicationGroupOperation}
-    return identity_element!(G, q)
-end
+_inv!(G::SpecialOrthogonalGroup, k, g) = copyto!(G, k, transpose(g))
 
 @doc "$(_doc_log_O2_id)"
 ManifoldsBase.log(::SpecialOrthogonalGroup{ManifoldsBase.TypeParameter{Tuple{2}}}, g)