+++ src/groups/unitary_group.jl
@@ -75,12 +75,8 @@ function identity_element(
     return fill(identity_element(G, T), 1, 1)
 end
 
-Base.inv(::UnitaryGroup, g) = adjoint(g)
-Base.inv(::UnitaryGroup, g::Identity{MatrixMultiplicationGroupOperation}) = g
-inv!(G::UnitaryGroup, h, g) = copyto!(G, h, adjoint(g))
-function inv!(G::UnitaryGroup, h, ::Identity{MatrixMultiplicationGroupOperation})
-    return identity_element!(G, h)
-end
+_inv(::UnitaryGroup, g) = adjoint(g)
+_inv!(G::UnitaryGroup, h, g) = copyto!(G, h, adjoint(g))
 
 function Base.log(
         ::UnitaryGroup{ManifoldsBase.ℍ, ManifoldsBase.TypeParameter{Tuple{1}}}, g::Number